Documentation of metadata/listobjectheaders for sub-type was incorrect
Since this is a customer-facing error, can we merge into the next release ASAP and also backport to get it corrected in production? Martha Miller had a customer failing when attempting to implement the metadata/listobjectheaders endpoint from this documentation. Current release version does not show that the 'subtypes' argument needs to be a JSON array (even in Swagger, it only says 'list of' as the hint). I've updated the documentation to mention the JSON array and updated the cURL and URL examples.